notification_service/requirements.txt
Django==5.0
djangorestframework
celery==5.3
redis